Quick (and dirty) fix of OMP race condition in polarisable code
This commit partly fixes #60 but it's dirty - single and multithreaded results are consistent
but without any performance gain. This is given mainly by the weak coupling between
threads, i.e. every thread has its own copy of math::VArray, which is reduced after every
iteration. It should be rewritten so the threads share the variables as much as possible.