Improved error messages on CUDA related code
GROMOS now tries to initialize GPU for SHAKE and/or INNERLOOP. If it fails, it returns the ID of uninitialized GPU. It also now allows -1 value for automatic decision and complains more, if user messes up the CUDA part of the INNERLOOP or CONSTRAINT block.

In blockinput.h, allow_missing was modified such that the caller knows, if the parameter was found or not. It is then responsibility of the caller to block.reset_error().