Test of precision in casting precedence
Several parameters are pre-calculated for GPU (e.g. squared cutoffs, inverted boxes, etc.). These have to be transferred to GPU in a single precision (float). One can calculate these values in double precision and then cast to float, or cast first into single precision. The latter is less precise (and here I found it is about 2-4x) and sadly that is how it was done so far.
From now on, I replace all operations with double precision and cast only at the latest moment. This leads to slightly different simulation results. I keep the float operations commented for the final code validation.