Deep CUDA implementation
This was a try to modify all basic types of the host code such that they can be also used in GPU code, but I discontinue this due to extreme complexity and nightmarish maintanability. Anyway, we will need to use a lot of stuff differently such that it is more efficient on CUDA, so sharing types between host and device is really undesirable.